Generate figures from the EUROSTAT quality report

This function uses the method iterate to create figures for a quality report for each component of a hierarchical persephone object

generateQrTable(x)

Arguments

x

a persephone object

Columns

component is a column added by iterate() and informs where in the hierarchy a certain component is.

Method is TS for tramoseats() and x13 for x13()

Period is 12 for yearly measurements and 4 for quarterly measurements

Start denotes the start of the time series

End denotes the end of the time series

Log.Transformation

ARIMA.Model

LeapYear

MovingHoliday

NbTF

Noutliers Number of outliers

Outlier1

Outlier2

Outlier3

CombinedTest_SI

Residual.Seasonality

Residual.TD.Effect

Q.Stat

Final.Henderson.Filter

Stage.2.Henderson.Filter

Seasonal.Filter

Max.Adj

Examples

obj_x13 <- perX13(AirPassengers, "RSA3")
ht <- perHts(a = obj_x13, b = obj_x13, method = "x13")
ht$run()
generateQrTable(ht)
#>   component Method Period Nobs  Start     End Log.Transformation    ARIMA.Model
#> 1              x13     12  144 1-1949 12-1960               TRUE (0 1 0)(0 1 1)
#> 2         a    x13     12  144 1-1949 12-1960               TRUE (0 1 1)(0 1 1)
#> 3         b    x13     12  144 1-1949 12-1960               TRUE (0 1 1)(0 1 1)
#>   LeapYear MovingHoliday NbTD Noutliers    Outlier1 Outlier2 Outlier3
#> 1    FALSE          TRUE    6         1 AO (5-1951)     <NA>     <NA>
#> 2    FALSE          TRUE    0         0        <NA>     <NA>     <NA>
#> 3    FALSE          TRUE    0         0        <NA>     <NA>     <NA>
#>   CombinedTest_SI Residual.Seasonality Residual.TD.Effect Q.Stat
#> 1         Present                   No                 No   0.22
#> 2         Present                   No                Yes   0.24
#> 3         Present                   No                Yes   0.24
#>   Final.Henderson.Filter Stage.2.Henderson.Filter Seasonal.Filter Max.Adj
#> 1                     H9                      H13             3x3     26%
#> 2                     H9                      H13             3x3     25%
#> 3                     H9                      H13             3x3     25%