Define a hierarchical time series

Combine mutliple objects of persephone objects into a new persephone object. The resulting time series can perform direct and indirect adjustments.

hierarchicalTimeSeries

Format

An object of class R6ClassGenerator of length 24.

Inherits

persephone

Constructor

hierarchicalTimeSeries$new(...)

... One or more persephone objects that use the same time instances.

Examples

obj_x13 <- x13Single$new(AirPassengers, "RSA3")

ht <- hierarchicalTimeSeries$new(obj_x13, obj_x13)
ht$run()
ht$adjusted

#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 1949 245.3656 249.3203 250.2444 254.9434 254.7441 252.3840 250.3420 253.4715
#> 1950 253.9774 267.5681 268.2003 265.7471 266.7281 278.2862 285.8888 290.1614
#> 1951 320.0953 320.7672 330.0688 331.1522 364.8111 327.9096 337.3442 339.8653
#> 1952 376.5488 376.6426 365.9700 365.4813 380.1928 398.1416 391.2097 403.4533
#> 1953 430.4578 434.4891 453.5157 472.7524 463.8361 444.5770 443.0021 450.4175
#> 1954 446.0889 424.7026 464.9303 459.7014 472.7064 478.1681 491.4836 485.7407
#> 1955 527.1542 533.3148 536.4656 543.6210 550.4755 565.4741 577.9376 575.1847
#> 1956 623.6428 618.5807 628.8582 646.4777 653.1191 659.7326 655.9927 662.7614
#> 1957 699.9198 705.1140 719.7466 720.4865 729.7838 734.4170 738.7302 744.1685
#> 1958 756.2739 751.0600 742.2385 725.7082 739.7200 764.8518 773.0967 787.0239
#> 1959 795.2041 813.4532 832.3737 836.9664 841.6383 839.0191 860.0234 867.7490
#> 1960 912.7280 903.1250 896.5981 942.2556 950.9364 950.0463 954.4892 959.7281
#>           Sep      Oct      Nov      Dec
#> 1949 257.7152 258.2676 263.3417 260.9519
#> 1950 296.9968 291.4951 289.3529 306.1854
#> 1951 345.0109 356.6952 368.0306 362.5473
#> 1952 401.8923 419.6740 424.3920 433.7900
#> 1953 454.3864 459.1703 450.0597 451.9866
#> 1954 494.6854 494.5850 512.7118 515.4811
#> 1955 594.2809 591.8258 598.8429 621.1969
#> 1956 661.7233 672.5127 684.5449 681.1317
#> 1957 757.8441 756.8868 760.6982 757.7076
#> 1958 765.8969 781.8622 766.9378 770.5602
#> 1959 878.0461 875.5946 901.2364 922.2459
#> 1960 962.9451 978.6772 978.9896 974.4346

ht$adjusted_indirect

#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 1949 249.0933 249.2524 249.7829 258.1414 250.2574 253.5105 250.5055 252.8224
#> 1950 254.3377 267.6473 266.3624 271.7894 257.6478 279.6893 287.7528 289.1336
#> 1951 318.2279 323.1674 335.9575 330.9310 352.1162 333.6342 336.6925 335.7959
#> 1952 373.4566 394.6386 366.8728 370.0723 372.2415 406.2476 386.6102 403.9181
#> 1953 427.8013 437.6921 454.3058 481.4087 463.6818 448.3528 438.4691 451.7676
#> 1954 445.9286 424.9672 461.1019 465.4770 474.9302 479.7461 493.1707 485.2342
#> 1955 529.3473 532.1794 532.4163 552.6928 550.9526 563.6328 587.6391 572.1776
#> 1956 622.5759 638.5643 640.0538 646.8851 653.4856 660.6607 660.2050 661.2192
#> 1957 692.7385 701.6014 723.1012 724.1610 730.2111 742.2598 738.2443 752.8172
#> 1958 750.4343 747.9669 739.3736 728.1694 743.6824 766.7991 771.4876 806.4644
#> 1959 795.3054 808.7360 833.7168 829.7870 853.1591 837.2460 855.7654 887.0882
#> 1960 921.0469 925.4952 866.0319 964.2515 953.5355 952.5890 967.0332 960.2308
#>           Sep      Oct      Nov      Dec
#> 1949 254.0016 260.0624 256.0928 258.3060
#> 1950 296.0704 290.3168 280.0139 306.6144
#> 1951 346.8315 351.3967 357.9969 364.0144
#> 1952 396.6189 411.9650 421.8049 427.0480
#> 1953 449.9785 453.0331 443.7474 444.3525
#> 1954 489.8932 493.3974 502.9211 508.4015
#> 1955 586.2457 593.0393 589.2241 618.1403
#> 1956 665.9338 664.2443 674.1866 682.3082
#> 1957 758.8772 753.1442 758.9071 751.5055
#> 1958 763.0615 777.0210 772.6965 757.6000
#> 1959 878.0997 879.5601 903.4672 913.7823
#> 1960 966.5037 993.5458 975.1511 977.8607

ht2 <- hierarchicalTimeSeries$new(ht, obj_x13)
ht2$run()
ht2$adjusted

#>            Jan       Feb       Mar       Apr       May       Jun       Jul
#> 1949  368.0484  373.9804  375.3665  382.4151  382.1161  378.5760  375.5130
#> 1950  380.9661  401.3522  402.3005  398.6207  400.0921  417.4292  428.8332
#> 1951  480.1429  481.1508  495.1033  496.7283  547.2167  491.8644  506.0162
#> 1952  564.8232  564.9639  548.9550  548.2220  570.2892  597.2124  586.8145
#> 1953  645.6867  651.7337  680.2735  709.1286  695.7542  666.8654  664.5031
#> 1954  669.1333  637.0539  697.3954  689.5520  709.0596  717.2522  737.2255
#> 1955  790.7313  799.9722  804.6984  815.4315  825.7132  848.2111  866.9064
#> 1956  935.4642  927.8711  943.2873  969.7165  979.6787  989.5989  983.9890
#> 1957 1049.8797 1057.6710 1079.6199 1080.7297 1094.6757 1101.6255 1108.0952
#> 1958 1134.4109 1126.5900 1113.3577 1088.5624 1109.5799 1147.2777 1159.6451
#> 1959 1192.8061 1220.1798 1248.5605 1255.4496 1262.4575 1258.5287 1290.0352
#> 1960 1369.0921 1354.6875 1344.8971 1413.3834 1426.4046 1425.0695 1431.7337
#>            Aug       Sep       Oct       Nov       Dec
#> 1949  380.2073  386.5728  387.4014  395.0126  391.4279
#> 1950  435.2420  445.4953  437.2426  434.0294  459.2781
#> 1951  509.7980  517.5164  535.0428  552.0458  543.8210
#> 1952  605.1799  602.8384  629.5110  636.5880  650.6849
#> 1953  675.6263  681.5795  688.7555  675.0896  677.9799
#> 1954  728.6111  742.0281  741.8775  769.0677  773.2217
#> 1955  862.7771  891.4214  887.7386  898.2644  931.7954
#> 1956  994.1421  992.5849 1008.7690 1026.8174 1021.6975
#> 1957 1116.2527 1136.7662 1135.3302 1141.0473 1136.5615
#> 1958 1180.5358 1148.8453 1172.7933 1150.4067 1155.8403
#> 1959 1301.6234 1317.0691 1313.3918 1351.8545 1383.3689
#> 1960 1439.5921 1444.4177 1468.0158 1468.4845 1461.6519

ht2$adjusted_indirect

#>            Jan       Feb       Mar       Apr       May       Jun       Jul
#> 1949  369.9123  373.9464  375.1358  384.0141  379.8728  379.1392  375.5947
#> 1950  381.1463  401.3918  401.3815  401.6418  395.5520  418.1308  429.7652
#> 1951  479.2093  482.3509  498.0476  496.6177  540.8692  494.7267  505.6904
#> 1952  563.2771  573.9619  549.4064  550.5174  566.3136  601.2654  584.5148
#> 1953  644.3585  653.3352  680.6686  713.4567  695.6770  668.7534  662.2366
#> 1954  669.0532  637.1862  695.4812  692.4399  710.1715  718.0412  738.0690
#> 1955  791.8278  799.4045  802.6738  819.9674  825.9518  847.2905  871.7571
#> 1956  934.9307  937.8628  948.8851  969.9202  979.8620  990.0630  986.0952
#> 1957 1046.2890 1055.9147 1081.2972 1082.5670 1094.8894 1105.5469 1107.8523
#> 1958 1131.4910 1125.0434 1111.9253 1089.7929 1111.5612 1148.2513 1158.8405
#> 1959 1192.8568 1217.8212 1249.2321 1251.8599 1268.2179 1257.6421 1287.9062
#> 1960 1373.2515 1365.8726 1329.6140 1424.3813 1427.7042 1426.3408 1438.0057
#>            Aug       Sep       Oct       Nov       Dec
#> 1949  379.8827  384.7160  388.2988  391.3881  390.1049
#> 1950  434.7281  445.0320  436.6535  429.3599  459.4926
#> 1951  507.7633  518.4267  532.3936  547.0290  544.5545
#> 1952  605.4123  600.2017  625.6565  635.2945  647.3140
#> 1953  676.3013  679.3756  685.6868  671.9335  674.1629
#> 1954  728.3578  739.6320  741.2837  764.1723  769.6819
#> 1955  861.2735  887.4038  888.3454  893.4550  930.2671
#> 1956  993.3710  994.6902 1004.6348 1021.6382 1022.2858
#> 1957 1120.5771 1137.2827 1133.4589 1140.1517 1133.4604
#> 1958 1190.2560 1147.4276 1170.3727 1153.2861 1149.3602
#> 1959 1311.2931 1317.0960 1315.3746 1352.9700 1379.1371
#> 1960 1439.8434 1446.1969 1475.4501 1466.5652 1463.3649