Define a hierarchical time series

Combine mutliple objects of persephone objects into a new persephone object. The resulting time series can perform direct and indirect adjustments.

perHts(
  ...,
  method = c("tramoseats", "x13"),
  userdefined = NULL,
  spec = NULL,
  list = NULL,
  weights = NULL
)

Arguments

...

ne or more objects which are either of class persephone or can be coerced to persephone objects with as_persephone. If more than one element is supplied, the underlying time series must have the same time instances. All elements supplied in ... must be named.

method

specifies the method to be used for the direct adjustment of the aggregate series. tramoseats or x13

userdefined

passed as the userdefined argument to tramoseats() or x13()

spec

a model specification returned by x13_spec() or tramoseats_spec()

list

a list of persephone objects as alternative input to .... This argument can also handle mts objects

weights

either a vector if the same weight is used for all time points or a list of ts objects or a mts object if the weight varies for different time points. They must have the same length as the number of components.

Examples

if (FALSE) {
objX13 <- perX13(AirPassengers, "RSA3")

ht <- perHts(a = objX13, b = objX13, method = "x13")

ht$updateParams(easter.enabled = TRUE)

ht$updateParams(component = "a", usrdef.outliersEnabled = TRUE,
                     usrdef.outliersType = c("AO","LS","LS"),
                     usrdef.outliersDate=c("1950-01-01","1955-04-01","1959-10-01"))
ht$run()
ht$adjustedDirect
ht$adjustedIndirect

ht2 <- perHts(a = ht, b = objX13)
ht2$run()
ht2$adjustedDirect
ht2$adjustedIndirect
}