persephone class

Common interface for x13 and tramoseats

Constructor

new(): Initializes an object. See the documentation for the derived classes for details.

Methods

run(verbose = FALSE): Runs the model

updateParams(...): Updates the current parameters

plot(...): Interactive time series plot for a persephone object, see plot.persephoneSingle or plot.hierarchicalTimeSeries

plotSeasIrrCal(...): Interactive plot of the seasonal component, irregular component and calendar effects for a persephone object, see plotSeasIrrCal

plotResiduals(...): Several interactive plots in connection with residuals for a persephone object, see plotResiduals

Read-Only fields

params: A parameters object of class SA_spec. See x13_spec and tramoseats_spec.

ts: The time series.

output: The return value from the underlying functions x13 or tramoseats. This slot will be empty (NULL) before run() is called for the first time.

Parameters

verbose: should the output from the underlying RJDemetra function be displayed after the run? If FALSE (the default), the output will be returned invisibly.

derived classes

perX13(), perTramo(), perHts()

Active bindings

params

A parameters object of class SA_spec. See x13_spec() and tramoseats_spec().

ts

the (unajusted) time series

tsp

the tsp() of the underlying time series

output

The return value from the underlying functions x13 or tramoseats. This slot will be empty (NULL) before run() is called for the first time.

adjusted

get the adjusted series

adjustedDirect

see perHts

spec

specifications passed to x13() and tramoseats() when the $run() method is invoked

forecasts

get forecasts from the model

forecastsDirect

get direct forecasts from the model

Methods

Public methods

persephone$new()

persephone$run()

persephone$updateParams()

persephone$plot()

persephone$plotSeasIrrCal()

persephone$plotResiduals()

persephone$print()

persephone$iterate()

persephone$generateQrTable()

persephone$setOptions()

persephone$fixModel()

persephone$fixOutlier()

persephone$clone()

Method new()

create a new persephone object

Usage

persephone$new()

Details

this function should not be used directly because this class is abstract. Only initialize derived classes.

Method run()

perform a run of the model

Usage

persephone$run(verbose = FALSE)

Arguments

verbose

whether to show outputs

Method updateParams()

update parameters for the adjustment

Usage

persephone$updateParams(...)

Arguments

...

passed to x13_spec() of tramoseats_spec()

Method plot()

visualize the results of an adjustment

Usage

persephone$plot(...)

Arguments

...

Interactive time series plot for a persephone object, see plot.persephoneSingle() or plot.hierarchicalTimeSeries()

Method plotSeasIrrCal()

Interactive plot of the seasonal component, irregular component and calendar effects for a persephone object

Usage

persephone$plotSeasIrrCal(...)

Arguments

...

passed to plotSeasIrrCal()

Method plotResiduals()

visualize residuals

Usage

persephone$plotResiduals(...)

Arguments

...

passed to plotResiduals()

Method print()

display a persephone object in the console

Usage

persephone$print()

Method iterate()

see perHts

Usage

persephone$iterate(fun, asTable = FALSE, unnest = FALSE)

Arguments

fun

see perHts

asTable

see perHts

unnest

see perHts

Method generateQrTable()

create a table for the eurostat quality report

Usage

persephone$generateQrTable()

Method setOptions()

update options for the model

Usage

persephone$setOptions(userdefined = NA, spec = NA, recursive = TRUE)

Arguments

userdefined

see x13() and tramoseats()

spec

see x13() and tramoseats()

recursive

only applicable to hierarchical series. propagates the updates to sub-series. see perHts

Method fixModel()

fix the arima model

Usage

persephone$fixModel(verbose = FALSE)

Arguments

verbose

if TRUE the changed parameters will be reported

Method fixOutlier()

create a new single object

Usage

persephone$fixOutlier(timespan = 12, verbose = FALSE)

Arguments

timespan

number of months from the end of the time series where outliers are not fixed

verbose

if TRUE the changed parameters will be reported

Method clone()

The objects of this class are cloneable with this method.

Usage

persephone$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.