Interactive plot of the seasonal component, irregular component and calendar effects for a persephone object

Produces a dygraph (see the online documentation for more detail). The function generates an interactive time series plot of the seasonal component, irregular component and calendar effects for a persephone object together with one year forecasts

plotSeasIrrCal(
  x,
  main = NULL,
  forecasts = TRUE,
  rangeSelector = TRUE,
  drawPoints = FALSE,
  annualComparison = NULL
)

Arguments

x

an object of class persephone.

main

plot title

forecasts

logical flag indicating if forecasts should be plotted

rangeSelector

logical flag specifying if a range selector should be included in the plot

drawPoints

logical flag indicating if a small dot should be drawn at each point, in addition to a line going through the point.

annualComparison

integer corresponding to the month or quarter which should be highlighted in the plot for every year.

Value

Returns an object of class dygraphs.

Examples

data(AirPassengers, package = "datasets")
# Generate a persephone object, in this case an x13Single object
obj <- perX13(AirPassengers, "RSA1")

obj$run()
# Plot after run
plotSeasIrrCal(obj)