tramoseats class

an R6 class that represents a tramoseats model

Super classes

persephone::persephone -> persephone::persephoneSingle -> tramoseatsSingle

Methods

Public methods

tramoseatsSingle$clone()

Inherited methods

persephone::persephone$fixModel()

persephone::persephone$fixOutlier()

persephone::persephone$generateQrTable()

persephone::persephone$iterate()

persephone::persephone$plot()

persephone::persephone$plotResiduals()

persephone::persephone$plotSeasIrrCal()

persephone::persephone$print()

persephone::persephone$setOptions()

persephone::persephone$updateParams()

persephone::persephoneSingle$initialize()

persephone::persephoneSingle$run()

Method clone()

The objects of this class are cloneable with this method.

Usage

tramoseatsSingle$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

data(AirPassengers, package = "datasets")

obj <- perTramo(AirPassengers, "RSA3")
obj$run()
obj$ts
#>      Jan Feb Mar Apr May Jun Jul Aug Sep Oct Nov Dec
#> 1949 112 118 132 129 121 135 148 148 136 119 104 118
#> 1950 115 126 141 135 125 149 170 170 158 133 114 140
#> 1951 145 150 178 163 172 178 199 199 184 162 146 166
#> 1952 171 180 193 181 183 218 230 242 209 191 172 194
#> 1953 196 196 236 235 229 243 264 272 237 211 180 201
#> 1954 204 188 235 227 234 264 302 293 259 229 203 229
#> 1955 242 233 267 269 270 315 364 347 312 274 237 278
#> 1956 284 277 317 313 318 374 413 405 355 306 271 306
#> 1957 315 301 356 348 355 422 465 467 404 347 305 336
#> 1958 340 318 362 348 363 435 491 505 404 359 310 337
#> 1959 360 342 406 396 420 472 548 559 463 407 362 405
#> 1960 417 391 419 461 472 535 622 606 508 461 390 432