persephone Model Classes

persephone defines a number of model classes. Objects belonging to a model class can be passed to the main package functions for fitting and prediction.

Value

An S4 object of the appropriate class.

Details

Details on the PersephoneCumLink class: Model fitting is based on ordinal::clm(). Information on arguments formula, nominal, scale, link and threshold can be found there.

Slots

region

an object of class Region. A region of interest. See package cronus for more details

data

data.frame. The data on which the model will be fitted.

model

ANY. An object holding information about the fitting.

fitted

data.frame. The fitted data.

metrics

list. A list of metrics calculated using persephone::crossval().

formula

character. The model formula. See details.

nominal

character. The nominal effects. See details.

scale

character. The scale effects. See details.

link

character. The link function used.

threshold

character. The type of thresholds used. See details.

See also

Examples

if (FALSE) {
# Create a Region object
library(cronus)
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

# Create a model
object1 <- new("PersephoneQuasiBin",
               region = region,
               crop = "Corn",
               data = progress_ne$Corn,
               formula = "CumPercentage ~ Time + agdd") # PersephoneModel

# Create another model
object2 <- new("PersephoneCumLink",
               region = region,
               crop = "Soybeans",
               data = progress_ne$Soybeans,
               formula = "Stage ~ Time + agdd + adayl") # PersephoneModel

# Concatenate the models
object <- c(object1, object2) # PersephoneModelList
}