Define a hierarchical time series

Combine mutliple objects of persephone objects into a new persephone object. The resulting time series can perform direct and indirect adjustments.

hierarchicalTimeSeries

Format

An object of class R6ClassGenerator of length 24.

Inherits

persephone

Constructor

hierarchicalTimeSeries$new(...)

... should contain one or more persephone objects that use the same time instances. All lements supplied here must be named.

model specifies a model to be used. tramoseats or x13

userdefined is passed as the userdefined argument to tramoseats or x13

spec a model specification returned by x13_spec() or tramoseats_spec()

Examples

obj_x13 <- x13Single$new(AirPassengers, "RSA3")

ht <- hierarchicalTimeSeries$new(a = obj_x13, b = obj_x13)
ht$run()
ht$adjusted

#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 1949 247.7036 250.1979 250.3748 255.4177 252.0618 251.6617 251.0775 252.5168
#> 1950 254.1406 268.1937 267.8904 268.6198 259.4133 277.2898 287.7509 289.2288
#> 1951 318.7270 321.3227 331.9352 332.6846 353.1479 329.4955 336.8187 337.5953
#> 1952 374.7407 391.1307 370.8223 363.2247 373.6626 400.4818 387.3252 407.3480
#> 1953 430.3966 435.7648 455.1258 471.9575 463.9653 442.1652 438.3587 453.7076
#> 1954 448.8332 427.4309 461.7183 457.2463 474.0018 475.9133 491.8988 485.0362
#> 1955 531.0149 533.7084 531.9915 545.6180 548.7504 562.7389 584.7571 570.0377
#> 1956 623.5945 637.8969 627.5342 650.2606 647.2164 662.5028 657.8353 657.2562
#> 1957 693.9513 699.2021 725.3510 716.1934 723.0861 743.9910 734.8050 745.4283
#> 1958 749.8479 743.9347 746.9698 720.5072 738.5865 767.5748 770.4128 796.2286
#> 1959 794.2589 802.6158 829.6964 834.3093 849.3363 835.0441 855.6613 878.2767
#> 1960 918.9252 919.1232 885.2691 947.2621 950.9915 947.5861 967.8332 953.1770
#>           Sep      Oct      Nov      Dec
#> 1949 256.4804 260.2968 262.6824 260.0671
#> 1950 298.5281 291.4918 288.2290 308.8182
#> 1951 349.8270 354.5031 366.9385 366.6403
#> 1952 399.0670 416.6650 430.6291 430.4260
#> 1953 451.7611 459.1141 451.0863 448.4450
#> 1954 491.9079 498.2608 509.4776 511.7643
#> 1955 590.8912 597.0005 595.6679 622.5493
#> 1956 670.6225 666.5912 680.3918 688.3501
#> 1957 761.4188 752.5708 762.8242 759.0768
#> 1958 763.1984 776.7669 775.4727 765.0662
#> 1959 876.7375 877.5134 904.7689 919.1497
#> 1960 963.1056 990.5096 975.4786 981.1770

ht$adjusted_indirect

#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 1949 249.0933 249.2524 249.7829 258.1414 250.2574 253.5105 250.5055 252.8224
#> 1950 254.3377 267.6473 266.3624 271.7894 257.6478 279.6893 287.7528 289.1336
#> 1951 318.2279 323.1674 335.9575 330.9310 352.1162 333.6342 336.6925 335.7959
#> 1952 373.4566 394.6386 366.8728 370.0723 372.2415 406.2476 386.6102 403.9181
#> 1953 427.8013 437.6921 454.3058 481.4087 463.6818 448.3528 438.4691 451.7676
#> 1954 445.9286 424.9672 461.1019 465.4770 474.9302 479.7461 493.1707 485.2342
#> 1955 529.3473 532.1794 532.4163 552.6928 550.9526 563.6328 587.6391 572.1776
#> 1956 622.5759 638.5643 640.0538 646.8851 653.4856 660.6607 660.2050 661.2192
#> 1957 692.7385 701.6014 723.1012 724.1610 730.2111 742.2598 738.2443 752.8172
#> 1958 750.4343 747.9669 739.3736 728.1694 743.6824 766.7991 771.4876 806.4644
#> 1959 795.3054 808.7360 833.7168 829.7870 853.1591 837.2460 855.7654 887.0882
#> 1960 921.0469 925.4952 866.0319 964.2515 953.5355 952.5890 967.0332 960.2308
#>           Sep      Oct      Nov      Dec
#> 1949 254.0016 260.0624 256.0928 258.3060
#> 1950 296.0704 290.3168 280.0139 306.6144
#> 1951 346.8315 351.3967 357.9969 364.0144
#> 1952 396.6189 411.9650 421.8049 427.0480
#> 1953 449.9785 453.0331 443.7474 444.3525
#> 1954 489.8932 493.3974 502.9211 508.4015
#> 1955 586.2457 593.0393 589.2241 618.1403
#> 1956 665.9338 664.2443 674.1866 682.3082
#> 1957 758.8772 753.1442 758.9071 751.5055
#> 1958 763.0615 777.0210 772.6965 757.6000
#> 1959 878.0997 879.5601 903.4672 913.7823
#> 1960 966.5037 993.5458 975.1511 977.8607

ht2 <- hierarchicalTimeSeries$new(a = ht, b = obj_x13)
ht2$run()
ht2$adjusted

#>            Jan       Feb       Mar       Apr       May       Jun       Jul
#> 1949  371.5554  375.2969  375.5622  383.1265  378.0928  377.4925  376.6163
#> 1950  381.2110  402.2906  401.8356  402.9297  389.1200  415.9347  431.6263
#> 1951  478.0904  481.9840  497.9028  499.0269  529.7218  494.2433  505.2280
#> 1952  562.1111  586.6960  556.2335  544.8371  560.4939  600.7226  580.9878
#> 1953  645.5949  653.6472  682.6887  707.9363  695.9480  663.2478  657.5381
#> 1954  673.2498  641.1464  692.5775  685.8694  711.0027  713.8699  737.8481
#> 1955  796.5223  800.5626  797.9872  818.4270  823.1256  844.1083  877.1357
#> 1956  935.3918  956.8453  941.3013  975.3909  970.8246  993.7542  986.7530
#> 1957 1040.9269 1048.8031 1088.0265 1074.2901 1084.6292 1115.9865 1102.2075
#> 1958 1124.7719 1115.9020 1120.4547 1080.7607 1107.8797 1151.3622 1155.6193
#> 1959 1191.3883 1203.9237 1244.5447 1251.4639 1274.0044 1252.5662 1283.4919
#> 1960 1378.3878 1378.6849 1327.9037 1420.8932 1426.4872 1421.3791 1451.7497
#>            Aug       Sep       Oct       Nov       Dec
#> 1949  378.7751  384.7205  390.4452  394.0236  390.1006
#> 1950  433.8432  447.7922  437.2376  432.3435  463.2272
#> 1951  506.3929  524.7405  531.7547  550.4077  549.9604
#> 1952  611.0219  598.6005  624.9975  645.9436  645.6390
#> 1953  680.5614  677.6416  688.6712  676.6294  672.6674
#> 1954  727.5543  737.8618  747.3911  764.2164  767.6464
#> 1955  855.0566  886.3368  895.5007  893.5019  933.8240
#> 1956  985.8844 1005.9337  999.8868 1020.5878 1032.5251
#> 1957 1118.1425 1142.1282 1128.8562 1144.2362 1138.6152
#> 1958 1194.3429 1144.7976 1165.1504 1163.2090 1147.5993
#> 1959 1317.4151 1315.1063 1316.2701 1357.1534 1378.7245
#> 1960 1429.7655 1444.6584 1485.7643 1463.2180 1471.7654

ht2$adjusted_indirect

#>            Jan       Feb       Mar       Apr       May       Jun       Jul
#> 1949  372.2502  374.8241  375.2662  384.4884  377.1906  378.4169  376.3302
#> 1950  381.3095  402.0174  401.0716  404.5145  388.2372  417.1344  431.6273
#> 1951  477.8409  482.9064  499.9139  498.1501  529.2060  496.3126  505.1649
#> 1952  561.4690  588.4500  554.2587  548.2609  559.7833  603.6056  580.6303
#> 1953  644.2973  654.6109  682.2787  712.6618  695.8062  666.3416  657.5933
#> 1954  671.7975  639.9145  692.2693  689.9848  711.4669  715.7863  738.4841
#> 1955  795.6885  799.7981  798.1996  821.9644  824.2267  844.5553  878.5767
#> 1956  934.8824  957.1790  947.5611  973.7032  973.9593  992.8332  987.9378
#> 1957 1040.3205 1050.0028 1086.9016 1078.2739 1088.1917 1115.1209 1103.9272
#> 1958 1125.0651 1117.9181 1116.6566 1084.5919 1110.4277 1150.9743 1156.1566
#> 1959 1191.9116 1206.9838 1246.5548 1249.2028 1275.9158 1253.6671 1283.5440
#> 1960 1379.4487 1381.8708 1318.2851 1429.3879 1427.7592 1423.8806 1451.3497
#>            Aug       Sep       Oct       Nov       Dec
#> 1949  378.9280  383.4812  390.3280  390.7288  389.2201
#> 1950  433.7956  446.5634  436.6501  428.2360  462.1254
#> 1951  505.4932  523.2427  530.2015  545.9369  548.6475
#> 1952  609.3070  597.3765  622.6475  641.5315  643.9500
#> 1953  679.5914  676.7503  685.6307  672.9600  670.6212
#> 1954  727.6533  736.8545  744.9595  760.9382  765.9650
#> 1955  856.1265  884.0140  893.5201  890.2800  931.6195
#> 1956  987.8659 1003.5894  998.7133 1017.4851 1029.5041
#> 1957 1121.8369 1140.8574 1129.1429 1142.2777 1134.8296
#> 1958 1199.4608 1144.7291 1165.2774 1161.8209 1143.8662
#> 1959 1321.8208 1315.7874 1317.2934 1356.5025 1376.0409
#> 1960 1433.2924 1446.3574 1487.2825 1463.0542 1470.1073