Define a hierarchical time series

Combine mutliple objects of persephone objects into a new persephone object. The resulting time series can perform direct and indirect adjustments.

Inherits

persephone

Constructor

per_hts(..., method = c("tramoseats", "x13"),
        userdefined = NULL, spec = NULL)

... should contain one or more persephone objects that use the same time instances. All elements supplied here must be named.

list a list of persephone objects as alternative input to ....

weights either a vector if the same weight is used for all time points or a list of ts objects or a mts object if the weight varies for different time points. They must have the same length as the number of components.

method specifies the method to be used. tramoseats or x13

userdefined is passed as the userdefined argument to tramoseats() or x13()

spec a model specification returned by x13_spec() or tramoseats_spec()

Fields

$components. A list of persephone objects.

$adjusted_indirect. Results from indirect adjustments which means the components are first adjusted and the adjusted series are then aggregated

Methods

$iterate(fun, ...) can be used to iterate over the hierarchy tree. See iterate().

$set_options(userdefined, spec, recursive = TRUE, component = "") sets options for all entries of the dependency tree recursively if recursive = TRUE (the default). See vignette("persephone-hierarchical").

Examples

obj_x13 <- per_x13(AirPassengers, "RSA3")

ht <- per_hts(a = obj_x13, b = obj_x13, method = "x13")
ht$run()
ht$adjusted_direct

#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 1949 247.4333 250.5063 251.8663 256.3079 258.0207 253.7141 247.8066 251.5403
#> 1950 256.2113 267.9867 266.4155 268.0955 268.4155 277.8872 285.2607 290.0129
#> 1951 324.8560 321.9177 330.5563 332.5670 367.5066 328.4363 335.1135 335.2647
#> 1952 384.2115 376.7993 365.8045 370.9622 379.8490 399.6288 388.2427 399.3339
#> 1953 433.0595 436.4639 456.6316 482.7892 463.3978 448.6915 434.2227 450.2359
#> 1954 444.3139 426.7972 466.0530 465.3934 473.9089 481.0764 487.8139 487.0403
#> 1955 527.5084 537.3563 540.3207 547.7267 552.4866 567.0259 573.9156 577.3508
#> 1956 623.2465 624.1541 630.9512 650.8605 660.7935 658.2673 657.4647 660.2502
#> 1957 700.1314 707.0794 720.1027 723.5888 730.7491 738.4547 742.8244 748.7361
#> 1958 750.3780 752.0901 742.9718 723.9404 745.3778 764.2811 783.0925 789.9832
#> 1959 795.6599 813.4225 831.2343 836.8935 848.7265 844.1597 860.1836 873.0110
#> 1960 913.6072 903.9192 888.6081 942.2535 966.8491 954.6949 957.5465 961.8190
#>           Sep      Oct      Nov      Dec
#> 1949 254.0698 256.7592 257.1790 258.7676
#> 1950 293.8013 289.1435 281.3111 302.9529
#> 1951 342.8506 353.0475 358.8656 361.1036
#> 1952 398.8721 411.9258 419.2342 430.5736
#> 1953 451.7713 453.7287 442.8956 448.7393
#> 1954 491.4584 488.9696 508.8649 508.2885
#> 1955 587.2795 589.2365 596.2170 617.6481
#> 1956 660.5664 669.0413 677.4343 680.2099
#> 1957 754.5300 757.9904 757.4566 758.3358
#> 1958 767.2220 771.9548 766.7476 772.0778
#> 1959 879.8532 871.3135 897.8527 932.5017
#> 1960 960.1101 978.8231 980.7512 979.2626

ht$adjusted_indirect

#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 1949 249.0933 249.2524 249.7829 258.1414 250.2574 253.5105 250.5055 252.8224
#> 1950 254.3377 267.6473 266.3624 271.7894 257.6478 279.6893 287.7528 289.1336
#> 1951 318.2279 323.1674 335.9575 330.9310 352.1162 333.6342 336.6925 335.7959
#> 1952 373.4566 394.6386 366.8728 370.0723 372.2415 406.2476 386.6102 403.9181
#> 1953 427.8013 437.6921 454.3058 481.4087 463.6818 448.3528 438.4691 451.7676
#> 1954 445.9286 424.9672 461.1019 465.4770 474.9302 479.7461 493.1707 485.2342
#> 1955 529.3473 532.1794 532.4163 552.6928 550.9526 563.6328 587.6391 572.1776
#> 1956 622.5759 638.5643 640.0538 646.8851 653.4856 660.6607 660.2050 661.2192
#> 1957 692.7385 701.6014 723.1012 724.1610 730.2111 742.2598 738.2443 752.8172
#> 1958 750.4343 747.9669 739.3736 728.1694 743.6824 766.7991 771.4876 806.4644
#> 1959 795.3054 808.7360 833.7168 829.7870 853.1591 837.2460 855.7654 887.0882
#> 1960 921.0469 925.4952 866.0319 964.2515 953.5355 952.5890 967.0332 960.2308
#>           Sep      Oct      Nov      Dec
#> 1949 254.0016 260.0624 256.0928 258.3060
#> 1950 296.0704 290.3168 280.0139 306.6144
#> 1951 346.8315 351.3967 357.9969 364.0144
#> 1952 396.6189 411.9650 421.8049 427.0480
#> 1953 449.9785 453.0331 443.7474 444.3525
#> 1954 489.8932 493.3974 502.9211 508.4015
#> 1955 586.2457 593.0393 589.2241 618.1403
#> 1956 665.9338 664.2443 674.1866 682.3082
#> 1957 758.8772 753.1442 758.9071 751.5055
#> 1958 763.0615 777.0210 772.6965 757.6000
#> 1959 878.0997 879.5601 903.4672 913.7823
#> 1960 966.5037 993.5458 975.1511 977.8607

ht2 <- per_hts(a = ht, b = obj_x13)
ht2$run()
ht2$adjusted_direct

#>            Jan       Feb       Mar       Apr       May       Jun       Jul
#> 1949  371.5554  375.2969  375.5622  383.1265  378.0928  377.4925  376.6163
#> 1950  381.2110  402.2906  401.8356  402.9297  389.1200  415.9347  431.6263
#> 1951  478.0904  481.9840  497.9028  499.0269  529.7218  494.2433  505.2280
#> 1952  562.1111  586.6960  556.2335  544.8371  560.4939  600.7226  580.9878
#> 1953  645.5949  653.6472  682.6887  707.9363  695.9480  663.2478  657.5381
#> 1954  673.2498  641.1464  692.5775  685.8694  711.0027  713.8699  737.8481
#> 1955  796.5223  800.5626  797.9872  818.4270  823.1256  844.1083  877.1357
#> 1956  935.3918  956.8453  941.3013  975.3909  970.8246  993.7542  986.7530
#> 1957 1040.9269 1048.8031 1088.0265 1074.2901 1084.6292 1115.9865 1102.2075
#> 1958 1124.7719 1115.9020 1120.4547 1080.7607 1107.8797 1151.3622 1155.6193
#> 1959 1191.3883 1203.9237 1244.5447 1251.4639 1274.0044 1252.5662 1283.4919
#> 1960 1378.3878 1378.6849 1327.9037 1420.8932 1426.4872 1421.3791 1451.7497
#>            Aug       Sep       Oct       Nov       Dec
#> 1949  378.7751  384.7205  390.4452  394.0236  390.1006
#> 1950  433.8432  447.7922  437.2376  432.3435  463.2272
#> 1951  506.3929  524.7405  531.7547  550.4077  549.9604
#> 1952  611.0219  598.6005  624.9975  645.9436  645.6390
#> 1953  680.5614  677.6416  688.6712  676.6294  672.6674
#> 1954  727.5543  737.8618  747.3911  764.2164  767.6464
#> 1955  855.0566  886.3368  895.5007  893.5019  933.8240
#> 1956  985.8844 1005.9337  999.8868 1020.5878 1032.5251
#> 1957 1118.1425 1142.1282 1128.8562 1144.2362 1138.6152
#> 1958 1194.3429 1144.7976 1165.1504 1163.2090 1147.5993
#> 1959 1317.4151 1315.1063 1316.2701 1357.1534 1378.7245
#> 1960 1429.7655 1444.6584 1485.7643 1463.2180 1471.7654

ht2$adjusted_indirect

#>            Jan       Feb       Mar       Apr       May       Jun       Jul
#> 1949  373.6400  373.8785  374.6743  387.2121  375.3862  380.2657  375.7582
#> 1950  381.5066  401.4709  399.5435  407.6841  386.4717  419.5339  431.6292
#> 1951  477.3419  484.7512  503.9362  496.3965  528.1743  500.4513  505.0388
#> 1952  560.1849  591.9579  550.3092  555.1084  558.3623  609.3714  579.9153
#> 1953  641.7020  656.5382  681.4587  722.1130  695.5227  672.5292  657.7037
#> 1954  668.8930  637.4507  691.6528  698.2155  712.3953  719.6192  739.7561
#> 1955  794.0209  798.2690  798.6244  829.0392  826.4289  845.4492  881.4587
#> 1956  933.8638  957.8464  960.0807  970.3276  980.2285  990.9911  990.3075
#> 1957 1039.1077 1052.4022 1084.6519 1086.2415 1095.3167 1113.3897 1107.3664
#> 1958 1125.6514 1121.9503 1109.0604 1092.2541 1115.5236 1150.1986 1157.2313
#> 1959 1192.9581 1213.1041 1250.5751 1244.6805 1279.7386 1255.8690 1283.6481
#> 1960 1381.5703 1388.2428 1299.0479 1446.3772 1430.3033 1428.8835 1450.5497
#>            Aug       Sep       Oct       Nov       Dec
#> 1949  379.2336  381.0024  390.0935  384.1392  387.4590
#> 1950  433.7003  444.1056  435.4752  420.0209  459.9216
#> 1951  503.6938  520.2472  527.0951  536.9954  546.0216
#> 1952  605.8772  594.9284  617.9476  632.7074  640.5721
#> 1953  677.6514  674.9678  679.5496  665.6212  666.5288
#> 1954  727.8512  734.8399  740.0962  754.3817  762.6022
#> 1955  858.2664  879.3685  889.5590  883.8362  927.2105
#> 1956  991.8288  998.9007  996.3664 1011.2799 1023.4623
#> 1957 1129.2258 1138.3158 1129.7163 1138.3607 1127.2582
#> 1958 1209.6966 1144.5922 1165.5315 1159.0447 1136.4001
#> 1959 1330.6323 1317.1496 1319.3401 1355.2008 1370.6735
#> 1960 1440.3461 1449.7555 1490.3187 1462.7266 1466.7910

#-------- example with industrial price indices -----------

data(ipi_c_eu, package = "RJDemetra")
# Reducing the data set to the EU28 countries
ipi_eu <- ipi_c_eu[, -c(1:3, 32:37)]
# We now build persephone objects for the countries
# (the lowest level in the hierachy)
ts_28 <- lapply(ipi_eu, per_tramo)

# We want to add an extra layer and split the EU28 countries in two groups
ht_half_europe_1  <- per_hts(list = ts_28[1:14], method = "tramoseats")

# Alternative way to use do.call
ht_half_europe_2 <- do.call(per_hts, ts_28[15:28])

# Now we generate the object for EU28
ht_europe <- per_hts(
  halfEU_1 = ht_half_europe_1,
  halfEU_2 = ht_half_europe_2,
  spec = "RSA5c",
  method = "x13"
)

# start the seasonal adjustment
ht_europe$run()

# accessing the directly and indirectly adjusted series for EU28
ht_europe$adjusted_direct

#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 2000 2226.613 2256.290 2274.005 2286.720 2315.177 2329.826 2316.971 2327.037
#> 2001 2399.108 2382.473 2382.610 2326.333 2374.685 2372.302 2345.654 2371.401
#> 2002 2329.082 2360.312 2409.190 2380.533 2393.683 2389.707 2408.325 2406.297
#> 2003 2429.852 2430.534 2417.481 2441.071 2426.692 2431.195 2461.843 2462.501
#> 2004 2485.408 2522.640 2541.214 2573.599 2540.240 2549.567 2550.641 2532.228
#> 2005 2571.926 2557.727 2576.797 2593.479 2594.421 2606.515 2611.052 2611.124
#> 2006 2676.370 2686.157 2703.026 2710.112 2776.312 2758.799 2752.753 2787.133
#> 2007 2868.147 2896.950 2918.316 2866.961 2911.680 2904.036 2895.983 2896.843
#> 2008 2999.478 3029.301 2983.451 2986.436 2934.155 2916.008 2895.745 2863.786
#> 2009 2459.035 2453.532 2424.997 2421.387 2429.421 2419.915 2433.477 2428.334
#>           Sep      Oct      Nov      Dec
#> 2000 2347.050 2342.753 2367.073 2426.172
#> 2001 2352.058 2349.939 2332.311 2336.641
#> 2002 2410.678 2418.332 2433.016 2412.638
#> 2003 2458.402 2490.799 2497.509 2511.351
#> 2004 2559.088 2559.113 2548.426 2558.467
#> 2005 2646.552 2644.353 2681.754 2687.827
#> 2006 2788.669 2794.188 2804.150 2883.287
#> 2007 2868.807 2891.305 2910.260 2941.155
#> 2008 2820.048 2770.181 2656.196 2556.018
#> 2009 2437.857 2456.713 2475.633 2456.262

ht_europe$adjusted_indirect

#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 1990       NA       NA       NA       NA       NA       NA       NA       NA
#> 1991       NA       NA       NA       NA       NA       NA       NA       NA
#> 1992       NA       NA       NA       NA       NA       NA       NA       NA
#> 1993       NA       NA       NA       NA       NA       NA       NA       NA
#> 1994       NA       NA       NA       NA       NA       NA       NA       NA
#> 1995       NA       NA       NA       NA       NA       NA       NA       NA
#> 1996       NA       NA       NA       NA       NA       NA       NA       NA
#> 1997       NA       NA       NA       NA       NA       NA       NA       NA
#> 1998       NA       NA       NA       NA       NA       NA       NA       NA
#> 1999       NA       NA       NA       NA       NA       NA       NA       NA
#> 2000 2226.471 2262.388 2265.882 2289.933 2305.953 2320.142 2314.026 2340.976
#> 2001 2397.341 2384.596 2375.448 2334.457 2366.467 2362.219 2342.526 2377.861
#> 2002 2332.466 2362.311 2390.813 2391.117 2383.389 2385.947 2400.949 2411.780
#> 2003 2432.198 2430.833 2416.858 2434.575 2421.453 2431.730 2456.395 2460.047
#> 2004 2491.939 2525.967 2536.819 2566.998 2537.320 2540.764 2547.501 2533.713
#> 2005 2580.455 2555.429 2567.389 2591.888 2589.626 2602.308 2610.091 2609.722
#> 2006 2683.022 2680.697 2695.134 2707.867 2763.510 2756.612 2764.431 2785.607
#> 2007 2863.568 2881.132 2905.275 2862.589 2897.397 2901.349 2910.538 2894.940
#> 2008 2989.965 3007.009 2960.711 2969.344 2920.918 2920.537 2905.778 2864.316
#> 2009 2455.791 2440.875 2407.908 2405.070 2412.069 2420.155 2434.737 2441.150
#> 2010       NA       NA       NA       NA       NA       NA       NA       NA
#> 2011       NA       NA       NA       NA       NA       NA       NA       NA
#> 2012       NA       NA       NA       NA       NA       NA       NA       NA
#> 2013       NA       NA       NA       NA       NA       NA       NA       NA
#> 2014       NA       NA       NA       NA       NA       NA       NA       NA
#> 2015       NA       NA       NA       NA       NA       NA       NA       NA
#> 2016       NA       NA       NA       NA       NA       NA       NA       NA
#> 2017       NA       NA       NA       NA       NA       NA       NA       NA
#>           Sep      Oct      Nov      Dec
#> 1990       NA       NA       NA       NA
#> 1991       NA       NA       NA       NA
#> 1992       NA       NA       NA       NA
#> 1993       NA       NA       NA       NA
#> 1994       NA       NA       NA       NA
#> 1995       NA       NA       NA       NA
#> 1996       NA       NA       NA       NA
#> 1997       NA       NA       NA       NA
#> 1998       NA       NA       NA       NA
#> 1999       NA       NA       NA       NA
#> 2000 2340.449 2350.477 2372.021 2417.111
#> 2001 2348.963 2353.844 2338.302 2340.554
#> 2002 2411.740 2421.372 2432.002 2418.690
#> 2003 2462.583 2489.005 2500.450 2514.993
#> 2004 2561.175 2556.391 2553.806 2561.496
#> 2005 2652.682 2650.568 2680.690 2690.013
#> 2006 2800.569 2805.091 2812.850 2877.743
#> 2007 2898.105 2903.216 2924.926 2943.019
#> 2008 2851.298 2780.974 2671.322 2553.962
#> 2009 2460.673 2473.641 2491.471 2463.535
#> 2010       NA       NA       NA       NA
#> 2011       NA       NA       NA       NA
#> 2012       NA       NA       NA       NA
#> 2013       NA       NA       NA       NA
#> 2014       NA       NA       NA       NA
#> 2015       NA       NA       NA       NA
#> 2016       NA       NA       NA       NA
#> 2017       NA       NA       NA       NA

# accessing the directly and indirectly adjusted series for half of Europe
ht_europe$components$halfEU_2$adjusted_direct

#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 2000 1049.790 1059.608 1061.476 1083.690 1086.986 1089.063 1090.856 1101.975
#> 2001 1123.511 1114.726 1114.029 1095.600 1116.013 1095.997 1093.243 1102.353
#> 2002 1087.737 1104.741 1123.844 1118.347 1120.475 1122.221 1126.556 1122.314
#> 2003 1135.540 1147.182 1134.937 1145.109 1137.538 1141.691 1158.338 1166.114
#> 2004 1177.320 1184.362 1190.671 1218.449 1196.242 1202.840 1197.096 1209.601
#> 2005 1225.645 1203.221 1212.874 1211.381 1217.599 1232.416 1237.905 1222.271
#> 2006 1279.909 1274.606 1278.635 1288.193 1312.104 1311.798 1332.308 1320.201
#> 2007 1369.884 1375.735 1386.500 1377.413 1387.934 1398.271 1397.843 1388.289
#> 2008 1442.971 1458.657 1428.725 1429.971 1414.369 1414.866 1407.652 1395.564
#> 2009 1177.775 1168.172 1162.759 1161.753 1160.335 1169.660 1188.178 1192.875
#> 2010 1205.373 1217.296 1241.872 1252.240 1279.290 1290.290 1286.046 1298.202
#> 2011 1347.893 1339.817 1325.344 1324.108 1334.704 1322.089 1313.968 1329.797
#> 2012 1337.622 1311.202 1338.515 1333.622 1326.697 1319.915 1338.749 1350.993
#> 2013 1319.610 1307.584 1319.018 1317.351 1311.398 1317.017 1320.841 1345.212
#> 2014 1337.207 1357.773 1350.041 1355.191 1346.835 1366.701 1367.880 1342.743
#> 2015 1375.218 1376.889 1389.947 1397.139 1395.879 1411.428 1416.486 1401.787
#> 2016 1408.388 1432.938 1419.813 1434.371 1429.903 1435.410 1426.310 1434.139
#> 2017 1458.912 1486.348 1494.016 1482.390 1509.142 1501.132 1495.169 1526.914
#>           Sep      Oct      Nov      Dec
#> 2000 1112.300 1111.787 1127.397 1140.930
#> 2001 1092.843 1105.279 1093.821 1095.801
#> 2002 1126.664 1135.639 1133.850 1128.164
#> 2003 1155.701 1175.573 1188.179 1182.805
#> 2004 1212.163 1215.644 1200.209 1213.934
#> 2005 1252.273 1261.495 1264.178 1272.205
#> 2006 1340.097 1340.496 1355.904 1368.757
#> 2007 1397.176 1387.664 1404.986 1417.704
#> 2008 1382.937 1354.153 1283.681 1198.946
#> 2009 1214.870 1219.985 1223.769 1210.603
#> 2010 1301.451 1312.525 1317.609 1306.914
#> 2011 1323.816 1328.362 1318.576 1316.257
#> 2012 1332.547 1329.049 1309.281 1318.090
#> 2013 1349.832 1334.139 1339.159 1344.739
#> 2014 1349.975 1357.928 1360.306 1355.253
#> 2015 1397.337 1420.188 1420.005 1403.835
#> 2016 1440.357 1442.743 1446.532 1468.308
#> 2017 1523.916 1524.964 1553.651 1565.441

ht_europe$components$halfEU_2$adjusted_indirect

#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 1990       NA       NA       NA       NA       NA       NA       NA       NA
#> 1991       NA       NA       NA       NA       NA       NA       NA       NA
#> 1992       NA       NA       NA       NA       NA       NA       NA       NA
#> 1993       NA       NA       NA       NA       NA       NA       NA       NA
#> 1994       NA       NA       NA       NA       NA       NA       NA       NA
#> 1995       NA       NA       NA       NA       NA       NA       NA       NA
#> 1996       NA       NA       NA       NA       NA       NA       NA       NA
#> 1997       NA       NA       NA       NA       NA       NA       NA       NA
#> 1998       NA       NA       NA       NA       NA       NA       NA       NA
#> 1999       NA       NA       NA       NA       NA       NA       NA       NA
#> 2000 1047.605 1065.067 1063.742 1077.716 1082.487 1086.959 1091.009 1101.766
#> 2001 1129.047 1118.294 1114.893 1088.517 1110.786 1097.554 1094.330 1103.183
#> 2002 1091.572 1108.408 1121.981 1119.500 1118.197 1117.022 1127.847 1123.118
#> 2003 1140.335 1150.273 1133.848 1142.761 1136.168 1134.872 1156.373 1167.232
#> 2004 1180.887 1186.199 1196.693 1215.205 1191.610 1202.667 1195.480 1207.071
#> 2005 1220.362 1205.879 1213.166 1219.528 1219.210 1224.353 1238.662 1225.856
#> 2006 1276.592 1275.247 1284.868 1286.137 1315.499 1309.764 1324.638 1323.363
#> 2007 1368.574 1373.544 1391.024 1373.860 1388.257 1401.663 1395.552 1391.043
#> 2008 1440.464 1460.605 1423.073 1438.625 1415.470 1410.326 1406.173 1397.953
#> 2009 1180.229 1169.346 1165.869 1159.328 1162.051 1171.990 1185.822 1186.104
#> 2010 1209.463 1219.628 1244.049 1255.012 1273.345 1295.060 1285.962 1292.520
#> 2011 1343.336 1341.719 1325.903 1329.299 1330.857 1319.666 1317.057 1325.438
#> 2012 1338.030 1313.693 1340.547 1328.644 1321.198 1326.546 1336.902 1345.048
#> 2013 1321.069 1310.119 1319.725 1322.102 1308.308 1316.055 1321.553 1340.876
#> 2014 1341.515 1358.436 1346.976 1356.745 1345.123 1362.429 1364.556 1342.761
#> 2015 1380.190 1376.993 1389.944 1388.986 1396.519 1412.488 1413.923 1395.752
#> 2016 1416.134 1429.699 1415.805 1439.401 1426.304 1429.981 1429.523 1433.918
#> 2017 1461.798 1484.897 1493.773 1476.374 1506.201 1502.179 1491.924 1524.285
#>           Sep      Oct      Nov      Dec
#> 1990       NA       NA       NA       NA
#> 1991       NA       NA       NA       NA
#> 1992       NA       NA       NA       NA
#> 1993       NA       NA       NA       NA
#> 1994       NA       NA       NA       NA
#> 1995       NA       NA       NA       NA
#> 1996       NA       NA       NA       NA
#> 1997       NA       NA       NA       NA
#> 1998       NA       NA       NA       NA
#> 1999       NA       NA       NA       NA
#> 2000 1114.594 1107.245 1123.001 1144.591
#> 2001 1089.561 1103.149 1093.405 1095.144
#> 2002 1122.877 1130.988 1136.016 1131.622
#> 2003 1157.794 1172.961 1184.140 1187.365
#> 2004 1207.684 1213.803 1200.375 1216.130
#> 2005 1252.521 1253.511 1267.513 1273.959
#> 2006 1344.359 1337.953 1352.329 1373.108
#> 2007 1393.601 1386.796 1408.257 1414.422
#> 2008 1382.785 1350.528 1284.274 1201.096
#> 2009 1215.917 1216.440 1222.239 1213.911
#> 2010 1295.540 1312.687 1321.055 1311.162
#> 2011 1324.260 1323.361 1324.012 1319.001
#> 2012 1330.417 1331.769 1313.403 1315.218
#> 2013 1345.025 1334.872 1346.652 1345.978
#> 2014 1351.932 1360.667 1360.901 1359.258
#> 2015 1400.566 1424.801 1418.083 1406.179
#> 2016 1441.553 1442.563 1453.104 1470.335
#> 2017 1530.987 1529.990 1551.697 1566.482

# accessing the adjusted series for a country
ht_europe$components$halfEU_2$components$AT$adjusted_direct

#>            Jan       Feb       Mar       Apr       May       Jun       Jul
#> 1996  50.90527  50.99849  51.13906  50.66368  52.16428  52.07320  52.27379
#> 1997  54.50998  54.23184  54.30778  55.51295  53.78102  55.16677  55.17382
#> 1998  59.69304  59.63744  60.47518  60.36968  61.35332  60.81851  61.01227
#> 1999  61.73333  60.95660  62.48621  62.87883  63.63905  63.98078  63.09919
#> 2000  65.56486  69.20320  68.89192  69.87918  71.61112  70.61045  71.49127
#> 2001  73.43845  74.33814  73.80445  72.27462  72.53916  72.40257  73.15068
#> 2002  71.64123  72.35970  70.83076  73.73797  73.26171  74.17063  73.24248
#> 2003  75.04438  73.40707  74.07646  74.49631  72.19770  73.34224  74.28415
#> 2004  77.06476  77.52449  77.51524  78.66189  78.30949  79.33491  80.27697
#> 2005  80.65267  80.90459  81.51590  82.20442  82.03923  82.93792  83.97000
#> 2006  85.70745  86.74100  86.09953  86.75332  88.34653  89.35229  90.60076
#> 2007  93.16508  94.76947  95.66806  93.65402  95.55214  94.13598  95.54666
#> 2008  98.44697  98.91550  99.44631  98.29152  97.94857  96.71776  95.21946
#> 2009  87.04012  84.39347  85.83160  81.46892  80.63934  81.34303  82.88747
#> 2010  83.33845  85.79002  86.88423  86.06906  89.36215  90.08095  89.99874
#> 2011  94.97893  95.05121  95.07963  95.09799  95.87142  95.74158  95.62540
#> 2012  94.82877  93.92358  95.14463  95.25511  96.90298  96.32696  96.19039
#> 2013  95.77368  94.71082  95.79402  97.20021  97.22068  95.77958  96.78709
#> 2014  99.11870  98.96343  97.02483  98.07117  96.14360 100.23076  97.36935
#> 2015  99.45647  99.69369 100.47635  99.38009  99.64374 100.59384 100.89810
#> 2016 103.51234 102.35885 102.22526 102.72949 102.40862 101.04149 101.17837
#> 2017 102.11349 105.21858 105.54065 105.83351 105.89366 106.41612 108.82152
#>            Aug       Sep       Oct       Nov       Dec
#> 1996  52.01782  53.18459  52.01599  52.84601  53.79968
#> 1997  57.17670  56.66752  56.48737  57.42491  60.04996
#> 1998  60.37360  62.80242  62.25209  61.38214  60.09554
#> 1999  64.69921  65.87669  67.20416  71.38379  66.66382
#> 2000  71.51024  71.68373  72.32664  75.12323  75.44150
#> 2001  73.12610  71.89317  73.89103  70.81196  71.01283
#> 2002  70.34707  72.69169  72.83138  72.61662  72.31464
#> 2003  74.52439  74.09413  74.38812  75.00771  75.90654
#> 2004  80.87877  80.96985  81.55298  81.91639  80.77093
#> 2005  83.06571  83.28079  85.08087  86.46712  83.40576
#> 2006  90.51434  90.62735  92.23168  92.82092  92.50081
#> 2007  96.12551  95.06855  96.78959  93.79644  97.02555
#> 2008  97.69185  97.12401  93.64032  89.30679  90.36033
#> 2009  82.18010  83.77792  85.48212  85.41021  86.12168
#> 2010  91.76224  92.44533  92.29060  94.01707  93.29584
#> 2011  96.34722  96.30318  95.48553  96.58621  96.26748
#> 2012  97.90311  95.56871  96.60165  95.47490  95.60104
#> 2013  97.39493  98.01015  96.40354  97.24113  97.01471
#> 2014  96.48647  97.52702  96.23488  95.94769  99.96150
#> 2015 100.59576  99.68759 100.84928 100.86443  98.46271
#> 2016 102.46745 102.24906 103.33371 104.21622 104.62973
#> 2017 108.65660 108.38206 109.24789 110.60799 111.67769