persephone class

Common interface for x13 and tramoseats

Constructor

new(): Initializes an object. See the documentation for the derived classes for details.

Methods

run(verbose = FALSE): Runs the model

updateParams(...): Updates the current parameters

plot(...): Interactive time series plot for a persephone object, see plot.persephoneSingle or plot.hierarchicalTimeSeries

plotSeasIrrCal(...): Interactive plot of the seasonal component, irregular component and calendar effects for a persephone object, see plotSeasIrrCal

plotResiduals(...): Several interactive plots in connection with residuals for a persephone object, see plotResiduals

Read-Only fields

params: A parameters object of class SA_spec. See x13_spec and tramoseats_spec.

ts: The time series.

output: The return value from the underlying functions x13 or tramoseats. This slot will be empty (NULL) before run() is called for the first time.

Parameters

verbose: should the output from the underlying RJDemetra function be displayed after the run? If FALSE (the default), the output will be returned invisibly.

See also