Interactive time series plot for a persephone object

Produces a dygraph (see the online documentation for more detail) for objects of class persephone. The function generates an interactive time series plot of the original series, the seasonally adjusted (SA) series and the trend as well as one year forecasts with prediction intervals of the original series. If no run has been performed on the persephone object, only the original time series is plotted.

# S3 method for persephone
plot(x, main = NULL, forecasts = TRUE,
  showOutliers = TRUE, rangeSelector = TRUE, drawPoints = FALSE,
  annualComparison = NULL, ...)

Arguments

x
an object of class persephone.

main
plot title

forecasts
logical flag indicating if forecasts should be plotted

showOutliers
logical flag specifying if outliers should be highlighted in the plot

rangeSelector
logical flag specifying if a range selector should be included in the plot

drawPoints
logical flag indicating if a small dot should be drawn at each point, in addition to a line going through the point.

annualComparison
integer corresponding to the month or quarter which should be highlighted in the plot for every year.

...
other plotting parameters to affect the plot. Not currently used.

Value

Returns an object of class dygraphs.

Examples

# Monthly data
data(AirPassengers, package = "datasets")
# Generate a persephone object, in this case an x13Single object
obj <- x13Single$new(AirPassengers, "RSA1")
# Plot before run of persephone object
plot(obj, drawPoints = TRUE)

#> Registered S3 method overwritten by 'xts':
#>   method     from
#>   as.zoo.xts zoo 

obj$run()
# Plot after run
plot(obj, drawPoints=TRUE)
# Update some parameters of the persephone object
# (randomly specify some outliers for illustrative purposes only)
obj$updateParams(usrdef.outliersEnabled = TRUE,
                usrdef.outliersType = c("AO","LS","LS"),
                usrdef.outliersDate=c("1950-01-01","1955-04-01",
                                      "1959-10-01"))
# Perform run to make updateParams take effect
obj$run()
plot(obj)

# Quarterly data
data(UKgas, package = "datasets")
obj2 <- x13Single$new(UKgas, "RSA3")
plot(obj2)
obj2$run()
plot(obj2)

# Generate a persephone object, in this case a tramoseatsSingle object
obj3 <- tramoseatsSingle$new(UKgas, "RSA3")
plot(obj3)
obj3$run()
plot(obj3)